+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.28.07] - 2021-12-12
+ - unix2dos command.
+   - Changed to print the message being converted.
 ## [0.28.06] - 2021-12-12
 ### Changed
  - wc command.